requirements.txt
asgiref==3.4.1
awesome-slugify==1.6.5
Babel==2.9.1
certifi==2021.10.8
charset-normalizer==2.0.10
Django==3.2.7
django-admin-sortable2==1.0.3
django-ckeditor==6.2.0
django-debug-toolbar==3.2.4
django-extra-views==0.13.0
django-haystack==3.1.1
django-js-asset==1.2.2
django-oscar==3.1
django-phonenumber-field==5.2.0
django-ranged-response==0.2.0
django-rosetta==0.9.7
django-simple-captcha==0.5.17
django-tables2==2.3.4
django-treebeard==4.4
django-widget-tweaks==1.4.8
factory-boy==3.2.0
Faker==8.12.2
gunicorn==20.1.0
idna==3.3
phonenumbers==8.12.31
Pillow==8.3.2
polib==1.1.1
psycopg2==2.9.1
purl==1.6
pycountry==20.7.3
pysolr==3.9.0
python-dateutil==2.8.2
python-decouple==3.4
pytz==2021.1
regex==2022.4.24
requests==2.27.1
six==1.16.0
sorl-thumbnail==12.6.3
sqlparse==0.4.2
text-unidecode==1.3
Unidecode==0.4.21
urllib3==1.26.8
